+++ gazebo/gui/model/ModelCreator.cc
@@ -114,6 +114,10 @@ ModelCreator::ModelCreator()
       boost::bind(&ModelCreator::ModelChanged, this)));
 
   this->connections.push_back(
+      gui::model::Events::ConnectOpenLinkInspector(
+      boost::bind(&ModelCreator::OpenInspector, this, _1)));
+
+  this->connections.push_back(
       gui::Events::ConnectAlignMode(
         boost::bind(&ModelCreator::OnAlignMode, this, _1, _2, _3, _4)));
 
@@ -928,12 +932,11 @@ void ModelCreator::CreateLinkFromSDF(sdf::ElementPtr _linkElem)
   {
     boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
     this->allLinks[linkName] = link;
+    gui::model::Events::linkInserted(linkName);
   }
 
   rendering::ScenePtr scene = link->linkVisual->GetScene();
   scene->AddVisual(link->linkVisual);
-
-  gui::model::Events::linkAdded(leafName);
   this->ModelChanged();
 }
 
@@ -957,6 +960,9 @@ void ModelCreator::RemoveLink(const std::string &_linkName)
   if (!link)
     return;
 
+  // Copy before reference is deleted.
+  std::string linkName(_linkName);
+
   rendering::ScenePtr scene = link->linkVisual->GetScene();
   for (auto &it : link->visuals)
   {
@@ -978,6 +984,7 @@ void ModelCreator::RemoveLink(const std::string &_linkName)
     this->allLinks.erase(_linkName);
     delete link;
   }
+  gui::model::Events::linkRemoved(linkName);
 
   std::string leafName = _linkName;
   size_t idx = _linkName.find_last_of("::");
@@ -1152,8 +1159,7 @@ void ModelCreator::Stop()
 {
   if (this->addLinkType != LINK_NONE && this->mouseVisual)
   {
-    for (unsigned int i = 0; i < this->mouseVisual->GetChildCount(); ++i)
-        this->RemoveLink(this->mouseVisual->GetName());
+    this->RemoveLink(this->mouseVisual->GetName());
     this->mouseVisual.reset();
     emit LinkAdded();
   }
@@ -1284,7 +1290,7 @@ bool ModelCreator::OnMouseRelease(const common::MouseEvent &_event)
     {
       LinkData *link = this->allLinks[this->mouseVisual->GetName()];
       link->SetPose(this->mouseVisual->GetWorldPose()-this->modelPose);
-      gui::model::Events::linkAdded(link->GetName());
+      gui::model::Events::linkInserted(this->mouseVisual->GetName());
     }
 
     // reset and return
@@ -1470,6 +1476,11 @@ void ModelCreator::OpenInspector(const std::string &_name)
 {
   boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
   LinkData *link = this->allLinks[_name];
+  if (!link)
+  {
+    gzerr << "Link [" << _name << "] not found." << std::endl;
+    return;
+  }
   link->SetPose(link->linkVisual->GetWorldPose()-this->modelPose);
   link->UpdateConfig();
   link->inspector->move(QCursor::pos());